# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  ok: boolean;
  message?: string;
}

export type ProdutoCategoria =
  (typeof ProdutoCategoria)[keyof typeof ProdutoCategoria];

export const ProdutoCategoria = {
  mercado: "mercado",
  cozinha: "cozinha",
} as const;

export interface Produto {
  id: number;
  codigo?: string | null;
  nome: string;
  marca?: string | null;
  categoria: ProdutoCategoria;
  preco: number;
  custo: number;
  estoque: number;
  estoque_min: number;
  unidade: string;
  ativo: boolean;
  criado_em: string;
}

export type CriarProdutoInputCategoria =
  (typeof CriarProdutoInputCategoria)[keyof typeof CriarProdutoInputCategoria];

export const CriarProdutoInputCategoria = {
  mercado: "mercado",
  cozinha: "cozinha",
} as const;

export interface CriarProdutoInput {
  codigo?: string | null;
  nome: string;
  marca?: string | null;
  categoria: CriarProdutoInputCategoria;
  preco: number;
  custo?: number;
  estoque?: number;
  estoque_min?: number;
  unidade?: string;
}

export type EditarProdutoInputCategoria =
  (typeof EditarProdutoInputCategoria)[keyof typeof EditarProdutoInputCategoria];

export const EditarProdutoInputCategoria = {
  mercado: "mercado",
  cozinha: "cozinha",
} as const;

export interface EditarProdutoInput {
  codigo?: string | null;
  nome?: string;
  marca?: string | null;
  categoria?: EditarProdutoInputCategoria;
  preco?: number;
  custo?: number;
  estoque?: number;
  estoque_min?: number;
  unidade?: string;
  ativo?: boolean;
}

export interface ItemVendaInput {
  produto_id: number;
  quantidade: number;
  preco_unit: number;
}

export type RegistrarVendaInputCategoria =
  (typeof RegistrarVendaInputCategoria)[keyof typeof RegistrarVendaInputCategoria];

export const RegistrarVendaInputCategoria = {
  mercado: "mercado",
  cozinha: "cozinha",
} as const;

export type RegistrarVendaInputPagamento =
  (typeof RegistrarVendaInputPagamento)[keyof typeof RegistrarVendaInputPagamento];

export const RegistrarVendaInputPagamento = {
  dinheiro: "dinheiro",
  pix: "pix",
  cartao: "cartao",
  fiado: "fiado",
} as const;

export interface RegistrarVendaInput {
  categoria: RegistrarVendaInputCategoria;
  desconto?: number;
  pagamento: RegistrarVendaInputPagamento;
  cliente_id?: number | null;
  observacao?: string | null;
  itens: ItemVendaInput[];
}

export interface ItemVenda {
  id: number;
  venda_id: number;
  produto_id: number;
  nome_snap: string;
  quantidade: number;
  preco_unit: number;
  subtotal: number;
}

export interface Venda {
  id: number;
  sessao_id?: number | null;
  categoria: string;
  total: number;
  desconto: number;
  pagamento: string;
  cliente_id?: number | null;
  cliente_nome?: string | null;
  observacao?: string | null;
  criado_em: string;
}

export interface ResumoVendas {
  total: number;
  total_dinheiro: number;
  total_pix: number;
  total_cartao: number;
  total_fiado: number;
  num_vendas: number;
  mercado: number;
  cozinha: number;
}

export type MovimentoEstoqueInputTipo =
  (typeof MovimentoEstoqueInputTipo)[keyof typeof MovimentoEstoqueInputTipo];

export const MovimentoEstoqueInputTipo = {
  entrada: "entrada",
  saida: "saida",
  ajuste: "ajuste",
} as const;

export interface MovimentoEstoqueInput {
  produto_id: number;
  tipo: MovimentoEstoqueInputTipo;
  quantidade: number;
  motivo?: string | null;
}

export interface MovimentoEstoque {
  id: number;
  produto_id: number;
  produto_nome?: string | null;
  tipo: string;
  quantidade: number;
  motivo?: string | null;
  criado_em: string;
}

export interface Cliente {
  id: number;
  nome: string;
  telefone?: string | null;
  observacao?: string | null;
  criado_em: string;
}

export interface CriarClienteInput {
  nome: string;
  telefone?: string | null;
  observacao?: string | null;
}

export interface FiadoItem {
  id: number;
  venda_id?: number | null;
  valor: number;
  pago: boolean;
  pago_em?: string | null;
  criado_em: string;
}

export interface ExtratoCliente {
  cliente: Cliente;
  fiados: FiadoItem[];
  total_aberto: number;
}

export interface PagarFiadoInput {
  valor: number;
}

export interface ResumoFiado {
  cliente: Cliente;
  total_aberto: number;
}

export type SessaoCaixaStatus =
  (typeof SessaoCaixaStatus)[keyof typeof SessaoCaixaStatus];

export const SessaoCaixaStatus = {
  aberto: "aberto",
  fechado: "fechado",
} as const;

export interface SessaoCaixa {
  id: number;
  aberto_em: string;
  fechado_em?: string | null;
  fundo_inicial: number;
  total_dinheiro: number;
  total_pix: number;
  total_cartao: number;
  total_fiado: number;
  total_sangria: number;
  status: SessaoCaixaStatus;
}

export interface StatusCaixa {
  aberto: boolean;
  sessao?: SessaoCaixa | null;
}

export interface AbrirCaixaInput {
  fundo_inicial?: number;
}

export interface Sangria {
  id: number;
  sessao_id: number;
  valor: number;
  motivo?: string | null;
  criado_em: string;
}

export interface SangriaInput {
  valor: number;
  motivo?: string | null;
}

export type ListarProdutosParams = {
  q?: string;
  categoria?: ListarProdutosCategoria;
};

export type ListarProdutosCategoria =
  (typeof ListarProdutosCategoria)[keyof typeof ListarProdutosCategoria];

export const ListarProdutosCategoria = {
  mercado: "mercado",
  cozinha: "cozinha",
} as const;

export type BuscarProdutoParams = {
  codigo?: string;
  nome?: string;
};

export type ListarVendasParams = {
  data?: string;
  categoria?: string;
  limit?: number;
};

export type ListarMovimentosParams = {
  produto_id?: number;
  limit?: number;
};

export type ListarClientesParams = {
  q?: string;
};

export type ListarSangriasParams = {
  sessao_id?: number;
};