# lib/api-zod/src/generated/types/criarProdutoInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */
import type { CriarProdutoInputCategoria } from "./criarProdutoInputCategoria";

export interface CriarProdutoInput {
  codigo?: string | null;
  nome: string;
  marca?: string | null;
  categoria: CriarProdutoInputCategoria;
  preco: number;
  custo?: number;
  estoque?: number;
  estoque_min?: number;
  unidade?: string;
}