# lib/api-zod/src/generated/types/editarProdutoInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */
import type { EditarProdutoInputCategoria } from "./editarProdutoInputCategoria";

export interface EditarProdutoInput {
  codigo?: string | null;
  nome?: string;
  marca?: string | null;
  categoria?: EditarProdutoInputCategoria;
  preco?: number;
  custo?: number;
  estoque?: number;
  estoque_min?: number;
  unidade?: string;
  ativo?: boolean;
}