# lib/api-zod/src/generated/types/extratoCliente.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */
import type { Cliente } from "./cliente";
import type { FiadoItem } from "./fiadoItem";

export interface ExtratoCliente {
  cliente: Cliente;
  fiados: FiadoItem[];
  total_aberto: number;
}