# lib/api-zod/src/generated/types/fiadoItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */

export interface FiadoItem {
  id: number;
  venda_id?: number | null;
  valor: number;
  pago: boolean;
  pago_em?: string | null;
  criado_em: string;
}