# lib/api-zod/src/generated/types/sangria.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */

export interface Sangria {
  id: number;
  sessao_id: number;
  valor: number;
  motivo?: string | null;
  criado_em: string;
}