# lib/api-zod/src/generated/types/sangriaInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */

export interface SangriaInput {
  valor: number;
  motivo?: string | null;
}