# lib/api-zod/src/generated/types/sessaoCaixa.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arvoredo PDV API
 * OpenAPI spec version: 0.1.0
 */
import type { SessaoCaixaStatus } from "./sessaoCaixaStatus";

export interface SessaoCaixa {
  id: number;
  aberto_em: string;
  fechado_em?: string | null;
  fundo_inicial: number;
  total_dinheiro: number;
  total_pix: number;
  total_cartao: number;
  total_fiado: number;
  total_sangria: number;
  status: SessaoCaixaStatus;
}